Added assisted leveling function and attempted temperature percentage
I have not tested this yet, but it should work in theory. I added a
function to help with leveling the print bed easily and quickly. There
is also an (attempted) method of getting the percentage from how hot
the printer is to how hot it should be when heating up for switching
filaments.